passt: Add new object to secondary objects and cache
In order to support reconcile in case one of the new objects is deleted / changed
(for example the daemonSet), add the new types to the watched objects list.

Note that NAD reconcile won't work if the CRD wasn't available in the cluster
once HCO is deployed.
It affects upstream but not OpenShift where multus is deployed before HCO by CNO.

Signed-off-by: Or Shoval <oshoval@redhat.com>